backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
sqlalchemy==2.0.38
alembic==1.14.1
python-multipart==0.0.20
gitpython==3.1.44
scikit-learn==1.6.1
numpy==2.2.3
tree-sitter==0.24.0
tree-sitter-language-pack==0.8.0
semgrep==1.108.0
libcst==1.6.0
pytest==8.3.4
httpx==0.28.1
pytest-asyncio==0.25.3
mypy==1.15.0
black==25.1.0
isort==6.0.0
ruff==0.9.6
